refactor(UserThemeController): SortOrder, SortColumn 매핑 Controller에서 하도록 변경
