+++ src/main/java/roomescape/theme/service/UserThemeService.java
@@ -5,6 +5,8 @@ import java.util.List;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import roomescape.theme.domain.Theme;
+import roomescape.theme.controller.SortColumn;
+import roomescape.theme.controller.SortOrder;
 import roomescape.theme.repository.ThemeRepository;
 
 @Service
@@ -17,8 +19,8 @@ public class UserThemeService {
     }
 
     @Transactional(readOnly = true)
-    public List<Theme> getThemes(String sort, String order, LocalDate startDate, LocalDate endDate, Long limit) {
-        return themeRepository.findRanked(sort, order, startDate, endDate, limit);
+    public List<Theme> getThemes(SortColumn sortColumn, SortOrder sortOrder, LocalDate startDate, LocalDate endDate, Long limit) {
+        return themeRepository.findRanked(sortColumn, sortOrder, startDate, endDate, limit);
     }
 
     @Transactional(readOnly = true)
